Move filter.js to lib/runtime/types
