+++ lib/runtime/types/index.js
@@ -3,7 +3,7 @@
 // Classes representing Juttle values which can't be represented by native
 // JavaScript types.
 
-var Filter = require('../filter');
+var Filter = require('./filter');
 var JuttleMoment = require('../../moment').JuttleMoment;
 
 module.exports = {
