+++ lib/runtime/values.js
@@ -4,7 +4,7 @@
 
 var _ = require('underscore');
 var ASTVisitor = require('../compiler/ast-visitor');
-var Filter = require('./filter');
+var Filter = require('./types/filter');
 var JuttleMoment = require('../moment').JuttleMoment;
 var errors = require('../errors');
