+++ test/runtime/values.spec.js
@@ -4,7 +4,7 @@ var expect = require('chai').expect;
 var _ = require('underscore');
 var values = require('../../lib/runtime/values');
 var JuttleMoment = require('../../lib/moment').JuttleMoment;
-var Filter = require('../../lib/runtime/filter');
+var Filter = require('../../lib/runtime/types/filter');
 
 describe('Values tests', function () {
     var FILTER_AST = {
