+++ test/spec/input-default-functions.js
@@ -2,7 +2,7 @@
 
 var _ = require('underscore');
 var JuttleMoment = require('../../lib/moment').JuttleMoment;
-var Filter = require('../../lib/runtime/filter');
+var Filter = require('../../lib/runtime/types/filter');
 
 function computeListDefault(input) {
     var items = input.options.items;
